Get numerical expressions expressioning
I think this is correct, but I guess I won't know for sure until I
create some kind of VM for running this and test the actual outputs
generated

The unary behaviour is ... questionable however the spec is very clear -
only the first value on a list can be negated and it can also be a `+`
if the programmer feels so inclined.

This in some ways is worse than the left associtivity of `^`.

Bonus points in that unary negation is the lowest precedence operator in
the entire stack