Add the DATA statement
This one was a complete bastard because the current state of my parser
doesn't let it backtrack far enough to parse `1a` as a string since `1`
is a valid number.

After being sad for a bit I realised that this is fine actually, because
the `DATA` statement doesn't know if the value will be parsed as a
string or a number when it's actually `READ` so keeping everything as
strings and attempting to do a numeric parse at runtime is much more
sensible - it would be at a minimum suprising if you had the program

```basic
1 DATA 1E1
2 READ $A
```

and `$A` contained `"10.0"` instead of `"1E1"`.

This does of course make doing runtime parsing future me's problem but
present me is absolutely fine with that